Refactor: Move all C sources to a new Runtime directory
This should help understand the structure of the source slightly better, as everything being dumped into a single folder with opaque naming conventions doesn't make for a great first impression.
