+++ Runtime/Bindings/lenv.c
@@ -15,7 +15,7 @@
  *
  */
 
-#include "./luvi.h"
+#include "../luvi.h"
 
 #ifdef __APPLE__
 #include <crt_externs.h>
